+++ CHANGELOG.md
@@ -6,9 +6,9 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
-- PVCs for data storage, cpu and memory limits are now configurable ([#241]).
+- PVCs for data storage, cpu and memory limits are now configurable ([#242]).
 
-[#241]: https://github.com/stackabletech/hive-operator/pull/241
+[#242]: https://github.com/stackabletech/hive-operator/pull/242
 
 ## [0.7.0] - 2022-09-06